refactor(styles): organize global and feature-based CSS structure
