+++ frontend/src/components/chatbox/ChatBox.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useRef, useState } from "react";
 import TypingIndicator from "./TypingIndicator";
 import OptionButton from "../common/OptionButton";
 import { useUserAnswers } from "../../context/UserAnswersContext";
-import "../../styles/ChatBox.css";
+import "../../styles/features/chatbox/ChatBox.css";
 
 export default function ChatBox({ messages = [], onSelect, collapse }) {
   const previousMessages = useRef([]);
