+++ frontend/src/styles/features/chatbox/TypingIndicator.css
