+++ frontend/src/styles/globals/background.css
@@ -30,16 +30,4 @@
     pointer-events: none;
   }
 
-  .scroll-hidden::-webkit-scrollbar {
-    display: none;
-  }
-  
-  .scroll-hidden {
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-  }
-  
-  .scroll-hidden {
-    scrollbar-gutter: stable both-edges;
-  }
