+++ frontend/src/styles/globals/scroll.css
@@ -1,0 +1,12 @@
+.scroll-hidden::-webkit-scrollbar {
+    display: none;
+  }
+  
+  .scroll-hidden {
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+  }
+  
+  .scroll-hidden {
+    scrollbar-gutter: stable both-edges;
+  }
